move builder to BillingWrapper, fix some docs
