+++ feature/google/src/main/java/com/revenuecat/purchases/google/billingClientParamBuilders.kt
@@ -52,50 +52,3 @@ val SubscriptionOfferDetails.subscriptionBillingPeriod: String?
 
 val SubscriptionOfferDetails.isBasePlan: Boolean
     get() = this.pricingPhases.pricingPhaseList.size == 1
-
-fun PurchaseOption.buildPurchaseParams(
-    storeProduct: StoreProduct,
-    replaceSkuInfo: ReplaceSkuInfo?,
-    appUserID: String
-): BillingFlowParams? {
-    val googlePurchaseOption = this as? GooglePurchaseOption
-    if (googlePurchaseOption == null) {
-        val error = PurchasesError(
-            PurchasesErrorCode.PurchaseInvalidError,
-            PurchaseStrings.INVALID_PURCHASE_OPTION_TYPE.format(
-                "Play",
-                "GooglePurchaseOption"
-            )
-        )
-        errorLog(error)
-        return null
-    }
-
-    val googleStoreProduct = storeProduct as? GoogleStoreProduct
-    if (googleStoreProduct == null) {
-        val error = PurchasesError(
-            PurchasesErrorCode.PurchaseInvalidError,
-            PurchaseStrings.INVALID_STORE_PRODUCT_TYPE.format(
-                "Play",
-                "GoogleStoreProduct"
-            )
-        )
-        errorLog(error)
-        return null
-    }
-
-    val productDetailsParamsList = BillingFlowParams.ProductDetailsParams.newBuilder().apply {
-        setOfferToken(googlePurchaseOption.token)
-        setProductDetails(storeProduct.productDetails)
-    }.build()
-
-    return BillingFlowParams.newBuilder()
-        .setProductDetailsParamsList(listOf(productDetailsParamsList))
-        .apply {
-            // only setObfuscatedAccountId for non-upgrade/downgrades until google issue is fixed:
-            // https://issuetracker.google.com/issues/155005449
-            replaceSkuInfo?.let {
-                setUpgradeInfo(it)
-            } ?: setObfuscatedAccountId(appUserID.sha256())
-        }.build()
-}
