+++ feature/google/src/main/java/com/revenuecat/purchases/google/purchaseOptionConversions.kt
@@ -2,8 +2,6 @@ package com.revenuecat.purchases.google
 
 import com.android.billingclient.api.ProductDetails
 import com.revenuecat.purchases.models.GooglePurchaseOption
-import com.revenuecat.purchases.models.GoogleStoreProduct
-import com.revenuecat.purchases.models.PurchaseOption
 
 fun ProductDetails.SubscriptionOfferDetails.toPurchaseOption(): GooglePurchaseOption {
     val pricingPhases = pricingPhases.pricingPhaseList.map { it.toRevenueCatPricingPhase() }
