+++ public/src/main/java/com/revenuecat/purchases/models/GooglePurchaseOption.kt
@@ -3,6 +3,9 @@ package com.revenuecat.purchases.models
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
+/**
+ * Defines an option for purchasing a Google subscription
+ */
 @Parcelize
 data class GooglePurchaseOption(
     override val pricingPhases: List<PricingPhase>,
